requirements.txt
streamlit==0.89.0
pytesseract==0.3.8
pillow==8.4.0
pandas==1.3.3
sqlite3==2.6.0
